test(cost): move tier sort regression to mapped suite
